Moved plugin to separate file
